feat(yl-bdf-orup): 续签token, 刷新token实现

+++ yl-bdf-orup/src/main/java/cc/yiueil/controller/OrupController.java
@@ -63,7 +63,7 @@ public class OrupController implements LoggedController {
         try {
             UserDto userDto = orupService.findUserByLoginName(loginName);
             if (password.equals(userDto.getPassword())) {
-                String token = JwtUtils.generateToken(userDto, ParseUtils.getInteger(GlobalProperties.getProperties("jwt.expire.seconds", "1800"), 1800));
+                String token = JwtUtils.generateToken(userDto, ParseUtils.getLong(GlobalProperties.getProperties("jwt.expire.seconds", "1800"), 1800L));
                 return success(userDto, "登录成功", token);
             }
         } catch (BusinessException e) {
@@ -83,7 +83,7 @@ public class OrupController implements LoggedController {
     public String refreshToken(HttpServletRequest request) {
         UserDto currentUser = getUser(request);
         UserDto user = orupService.getUser(currentUser.getId());
-        String newToken = JwtUtils.generateToken(user, ParseUtils.getInteger(GlobalProperties.getProperties("jwt.expire.seconds", "1800"), 1800));
+        String newToken = JwtUtils.generateToken(user, ParseUtils.getLong(GlobalProperties.getProperties("jwt.expire.seconds", "1800"), 1800L));
         return success(user, "刷新成功", newToken);
     }
